# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shop Backend API
  description: API completa para gestión de usuarios, productos, pedidos, carrito, pagos y notificaciones
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local
...
paths:
...
  /orders:
    post:
      summary: Crear un nuevo pedido
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Pedido creado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  orderId:
                    type: integer
        '400':
          description: Error en la creación del pedido

    get:
      summary: Listar todos los pedidos (admin)
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de todos los pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Error al obtener pedidos
  /orders/me:
    get:
      summary: Obtener todos los pedidos del usuario autenticado
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pedidos del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Error al obtener pedidos
  /orders/{id}:
    get:
      summary: Obtener un pedido por ID
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Pedido no encontrado

    delete:
      summary: Eliminar un pedido por ID (admin)
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido eliminado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error al eliminar pedido
  /orders/{id}/status:
    put:
      summary: Actualizar estado de un pedido (admin)
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '200':
          description: Estado actualizado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error al actualizar estado
  /orders/{id}/return:
    post:
      summary: Solicitar devolución de un pedido
      tags: [Returns]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestReturn'
      responses:
        '201':
          description: Devolución creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  returnId:
                    type: integer
        '400':
          description: Error al solicitar devolución
  /orders/me/returns:
    get:
      summary: Obtener todas las devoluciones del usuario
      tags: [Returns]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de devoluciones del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Return'
        '400':
          description: Error al obtener devoluciones
  /orders/returns/{id}/status:
    put:
      summary: Actualizar estado de una devolución (admin)
      tags: [Returns]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
      responses:
        '200':
          description: Estado de devolución actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error al actualizar devolución
...
  /payments/webhook:
    post:
      summary: Webhook de Stripe
      description: |
        Endpoint para recibir eventos de Stripe, como confirmaciones de pago (`checkout.session.completed`).
        Stripe valida este endpoint mediante la cabecera `stripe-signature`.
        Nota: este endpoint suele recibir el "raw body" de Stripe para validar la firma; documenta en tu cliente que debe usar raw body.
      tags: [Payments]
      parameters:
        - name: stripe-signature
          in: header
          required: true
          description: Firma enviada por Stripe para validar el evento
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Evento enviado por Stripe
      responses:
        '200':
          description: Webhook recibido correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    example: true
        '400':
          description: Error de validación de firma o payload inválido
        '500':
          description: Error interno procesando evento

  # -------------------- DASHBOARD --------------------
...
  /dashboard/orders:
    get:
      summary: Pedidos filtrados
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filtrar por estado del pedido
          schema:
            type: string
        - name: start
          in: query
          description: Fecha de inicio (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: end
          in: query
          description: Fecha de fin (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: userId
          in: query
          description: Filtrar por ID de usuario
          schema:
            type: integer
      responses:
        '200':
          description: Lista de pedidos filtrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FilteredOrder'
        '500':
          description: Error interno
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Address:
      type: object
      properties:
        id:
          type: integer
          example: 3
        first_name:
          type: string
          example: "María"
        last_name:
          type: string
          example: "Pérez"
        street:
          type: string
          example: "Calle Mayor 12, 3ºA"
        city:
          type: string
          example: "Madrid"
        province:
          type: string
          example: "Madrid"
        postal_code:
          type: string
          example: "28013"
        country:
          type: string
          example: "España"
        phone:
          type: string
          example: "+34 600 123 456"
        type:
          type: string
          example: "shipping"
        created_at:
          type: string
          format: date-time
          example: "2025-03-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-03-01T12:00:00Z"
...
    CreateOrderRequest:
      type: object
      required:
        - address_id
        - items
        - total
      properties:
        address_id:
          type: integer
        items:
          type: array
          items:
            type: object
            required:
              - product_id
              - quantity
              - price
            properties:
              product_id:
                type: integer
              quantity:
                type: integer
              price:
                type: number
        total:
          type: number
    Order:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        status:
          type: string
        status_label:
          type: string
        total:
          type: number
        total_paid:
          type: number
          nullable: true
        discount_amount:
          type: number
          nullable: true
        promotion_code:
          type: string
          nullable: true
        address:
          $ref: '#/components/schemas/Address'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        quantity:
          type: integer
        price:
          type: number
        image_url:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
    RequestReturn:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
    Return:
      type: object
      properties:
        id:
          type: integer
        order_id:
          type: integer
        user_id:
          type: integer
        reason:
          type: string
        total_amount:
          type: number
        status:
          type: string
        order_status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # -------------------- PAYMENTS --------------------
...
    FilteredOrder:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        total:
          type: number
          format: float
        status:
          type: string
        created_at:
          type: string
          format: date
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
...